+++ db/migrate/20150210142824_destroy_invalid_messages.rb
@@ -1,0 +1,5 @@
+class DestroyInvalidMessages < ActiveRecord::Migration
+  def change
+    Message.all.select { |m| m.recipient.nil? }.map(&:destroy!)
+  end
+end